requirements.txt
Django==3.2.8
django-cors-headers==3.10.0
djangorestframework==3.12.4
psycopg2==2.9.1
drf-yasg==1.20.0
PyJWT==2.2.0